requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2
django-cors-headers==3.8.0
djangorestframework==3.12.4
idna==3.3
Pillow==8.4.0
PyJWT==2.1.0
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
